+++ src/app/component-library/catalog/asset-detail-dialog/asset-detail-dialog.component.ts
@@ -10,6 +10,7 @@ import {filter, finalize, takeUntil} from 'rxjs/operators';
 import {UiContractOffer} from '@sovity.de/edc-client';
 import {MailtoLinkBuilder} from 'src/app/core/services/mailto-link-builder';
 import {EdcApiService} from '../../../core/services/api/edc-api.service';
+import {ConnectorLimitsService} from '../../../core/services/connector-limits.service';
 import {ContractNegotiationService} from '../../../core/services/contract-negotiation.service';
 import {UiAssetMapped} from '../../../core/services/models/ui-asset-mapped';
 import {NotificationService} from '../../../core/services/notification.service';
@@ -43,6 +44,8 @@ export class AssetDetailDialogComponent implements OnDestroy {
   data!: AssetDetailDialogData;
   asset!: UiAssetMapped;
   propGroups!: PropertyGridGroup[];
+
+  limitsExceeded: boolean | null = null;
 
   loading = false;
 
@@ -78,6 +81,7 @@ export class AssetDetailDialogComponent implements OnDestroy {
   constructor(
     private edcApiService: EdcApiService,
     private notificationService: NotificationService,
+    private connectorLimitsService: ConnectorLimitsService,
     private matDialog: MatDialog,
     private matDialogRef: MatDialogRef<AssetDetailDialogComponent>,
     @Inject(MAT_DIALOG_DATA)
@@ -99,6 +103,14 @@ export class AssetDetailDialogComponent implements OnDestroy {
     this.data = data;
     this.asset = this.data.asset;
     this.propGroups = this.data.propertyGridGroups;
+
+    if (this.limitsExceeded == null && this.data.type === 'data-offer') {
+      this.connectorLimitsService
+        .isConsumingAgreementLimitExceeded()
+        .subscribe((limitsExceeded) => {
+          this.limitsExceeded = limitsExceeded;
+        });
+    }
   }
 
   onContactClick() {
@@ -132,10 +144,22 @@ export class AssetDetailDialogComponent implements OnDestroy {
   }
 
   onNegotiateClick(contractOffer: UiContractOffer) {
-    this.contractNegotiationService.negotiate(
-      this.data.dataOffer!,
-      contractOffer,
-    );
+    this.connectorLimitsService
+      .isConsumingAgreementLimitExceeded()
+      .subscribe((limitExceeded) => {
+        if (!limitExceeded) {
+          this.limitsExceeded = false;
+          this.contractNegotiationService.negotiate(
+            this.data.dataOffer!,
+            contractOffer,
+          );
+        } else {
+          this.limitsExceeded = true;
+          this.notificationService.showError(
+            'Cannot negotiate. Maximum number of active consuming contracts reached.',
+          );
+        }
+      });
   }
 
   onTransferClick() {