+++ CHANGELOG.md
@@ -20,6 +20,8 @@ the detailed section referring to by linking pull requests or issues.
 
 #### Patch
 
+- Check contract limits before negotiating a new one.
+
 ## [v4.1.2] - 2024-08-20
 
 ### Overview